+++ iOS/Sources/Beagle/Sources/Components/Page+Renderable/ComponentHostController.swift
@@ -22,7 +22,7 @@ class ComponentHostController: BeagleController {
     let component: RawComponent
     let renderer: BeagleRenderer
 
-    var bindings: [() -> Void] = []
+    let bindings = Bindings()
 
     var dependencies: BeagleDependenciesProtocol {
         return renderer.controller.dependencies
@@ -47,27 +47,7 @@ class ComponentHostController: BeagleController {
     }
     
     public func addBinding<T: Decodable>(expression: ContextExpression, in view: UIView, update: @escaping (T?) -> Void) {
-         bindings.append { [weak self, weak view] in
-             guard let self = self else { return }
-             view?.configBinding(
-                 for: expression,
-                 completion: self.bindBlock(view: view, update: update)
-             )
-         }
-     }
-
-    private func bindBlock<T: Decodable>(view: UIView?, update: @escaping (T?) -> Void) -> (T?) -> Void {
-        return { [weak self, weak view] value in
-            update(value)
-            view?.yoga.markDirty()
-            self?.viewIfLoaded?.setNeedsLayout()
-        }
-    }
-
-    func configBindings() {
-        while let bind = bindings.popLast() {
-            bind()
-        }
+        bindings.add(self, expression, view, update)
     }
 
     init(_ component: RawComponent, renderer: BeagleRenderer) {
@@ -86,7 +66,7 @@ class ComponentHostController: BeagleController {
     }
 
     override func viewDidLayoutSubviews() {
-        configBindings()
+        bindings.config()
         dependencies.style(view).applyLayout()
         super.viewDidLayoutSubviews()
     }
