+++ iOS/Sources/Beagle/Sources/Navigation/Tests/BeagleNavigatorTests.swift
@@ -349,7 +349,7 @@ final class BeagleNavigatorTests: XCTestCase {
 
         // When
         view.setContext(Context(id: "url", value: "initial"))
-        controller.configBindings()
+        controller.bindings.config()
         setContext.execute(controller: controller, origin: view)
         sut.navigate(action: pushViewRemote, controller: controller, animated: false, origin: view)
